Remove helper.rs
