+++ examples/demo-stf/src/tests/tx_revert_tests.rs
@@ -1,9 +1,7 @@
 use core::panic;
 
 use crate::{
-    app::create_new_demo,
     config::{create_demo_config, LOCKED_AMOUNT, TEST_SEQUENCER_DA_ADDRESS},
-    helpers::new_test_blob,
     runtime::Runtime,
     tests::data_generation::simulate_da_with_bad_serialization,
 };
@@ -14,7 +12,11 @@ use sov_modules_api::{
 use sov_rollup_interface::{mocks::MockZkvm, stf::StateTransitionFunction};
 use sov_state::{ProverStorage, WorkingSet};
 
-use super::data_generation::{simulate_da_with_bad_sig, simulate_da_with_revert_msg};
+use super::{
+    create_new_demo,
+    data_generation::{simulate_da_with_bad_sig, simulate_da_with_revert_msg},
+    new_test_blob,
+};
 
 const SEQUENCER_BALANCE_DELTA: u64 = 1;
 const SEQUENCER_BALANCE: u64 = LOCKED_AMOUNT + SEQUENCER_BALANCE_DELTA;
